fix(ie-objects): move LimitAccessUserInfo interface to separate file
